mcp_a2a/pyproject.toml
[project]
name = "mcp-a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.6.1",
    "forex-python>=1.9.2",
    "google-adk>=1.1.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "lab>=8.4",
]
